Add population layer
A new child of the feature group, the population layer
changes the color of each country based on its population.

Also a minor fix for the user prompt for the starting location
of the map as the default location was hardcoded into the
map creation function thus user input was ignored.

GH-2
GH-5